679109 - localpath must be defined in except block
fixing
error: rpmts_HdrFromFdno: MD5 digest: BAD
Expected(d89999554a6046d7c9842d6fbe32e5a7) != (1a6542ee1f2e66a0c307ed2eefe83f84)
(50, 'Unable to load package', 'Invalid information uploaded to the server')
Traceback (most recent call last):
  File "/usr/bin/spacewalk-repo-sync", line 69, in ?
...
  File
"/usr/lib/python2.4/site-packages/spacewalk/satellite_tools/reposync.py", line
191, in import_packages
    finally_remove(localpath)
UnboundLocalError: local variable 'localpath' referenced before assignment

+++ backend/satellite_tools/reposync.py
@@ -168,11 +168,12 @@ class RepoSync:
         is_non_local_repo = (url.find("file://") < 0)
         # try/except/finally doesn't work in python 2.4 (RHEL5), so here's a hack
         def finally_remove(path):
-            if is_non_local_repo:
+            if is_non_local_repo and path:
                 os.remove(path)
 
         for (index, what) in enumerate(to_process):
             pack, to_download, to_link = what
+            localpath = None
             try:
                 self.print_msg("%d/%d : %s" % (index+1, num_to_process, pack.getNVREA()))
                 if to_download:
